# Conflicts:
#	tkpdseller/src/main/java/com/tokopedia/seller/product/view/activity/ProductAddActivity.java

Fix below
image picker error waktu compressed true
catalog picker blum seperti UI
set description image crash
category recommendation visibility
shop info remove param for deviceID loginId
shop transform - view model untuk addroductshopinfousecase

+++ tkpdseller/src/main/java/com/tokopedia/seller/product/data/mapper/CategoryServiceToDbMapper.java
@@ -36,7 +36,7 @@ public class CategoryServiceToDbMapper implements Func1<CategoryServiceModel, Li
         List<CategoryDataBase> dbModels = new ArrayList<>();
 
         CategoryDataBase dbModel = new CategoryDataBase();
-        dbModel.setId(Integer.parseInt(category.getId()));
+        dbModel.setId(Long.parseLong(category.getId()));
         dbModel.setIdentifier(category.getIdentifier());
         dbModel.setName(category.getName());
         dbModel.setParentId(parent);
