+++ tkpdseller/src/main/java/com/tokopedia/seller/product/data/repository/CatalogRepositoryImpl.java
@@ -1,8 +1,7 @@
 package com.tokopedia.seller.product.data.repository;
 
-import com.tokopedia.core.base.di.scope.ActivityScope;
-import com.tokopedia.seller.product.data.source.cloud.model.catalogdata.CatalogDataModel;
 import com.tokopedia.seller.product.data.source.CatalogDataSource;
+import com.tokopedia.seller.product.data.source.cloud.model.catalogdata.CatalogDataModel;
 import com.tokopedia.seller.product.domain.CatalogRepository;
 
 import javax.inject.Inject;
@@ -22,7 +21,7 @@ public class CatalogRepositoryImpl implements CatalogRepository {
     }
 
     @Override
-    public Observable<CatalogDataModel> fetchCatalog(String keyword, int prodDeptId, int start, int row) {
+    public Observable<CatalogDataModel> fetchCatalog(String keyword, long prodDeptId, int start, int row) {
         return catalogDataSource.fetchCatalog(keyword, prodDeptId, start, row);
     }
 }
