+++ tkpdseller/src/main/java/com/tokopedia/seller/product/domain/interactor/FetchCategoryFromSelectedUseCase.java
@@ -17,7 +17,7 @@ import rx.Observable;
  * @author sebastianuskh on 4/7/17.
  */
 
-public class FetchCategoryFromSelectedUseCase extends UseCase<List<CategoryLevelDomainModel>>{
+public class FetchCategoryFromSelectedUseCase extends UseCase<List<CategoryLevelDomainModel>> {
 
     public static final String INIT_SELECTED = "INIT_SELECTED";
     public static final int INIT_UNSELECTED = -1;
@@ -35,16 +35,16 @@ public class FetchCategoryFromSelectedUseCase extends UseCase<List<CategoryLevel
 
     @Override
     public Observable<List<CategoryLevelDomainModel>> createObservable(RequestParams requestParams) {
-        int initSelected = requestParams.getInt(INIT_SELECTED, INIT_UNSELECTED);
-        if (initSelected == INIT_UNSELECTED){
+        long initSelected = requestParams.getLong(INIT_SELECTED, INIT_UNSELECTED);
+        if (initSelected == INIT_UNSELECTED) {
             throw new RuntimeException("Init category is unselected");
         }
         return categoryRepository.fetchCategoryFromSelected(initSelected);
     }
 
-    public static RequestParams generateParam(int initSelected) {
+    public static RequestParams generateParam(long initSelected) {
         RequestParams requestParam = RequestParams.create();
-        requestParam.putInt(INIT_SELECTED, initSelected);
+        requestParam.putLong(INIT_SELECTED, initSelected);
         return requestParam;
     }
 }
