+++ tkpdseller/src/main/java/com/tokopedia/seller/product/view/adapter/category/CategoryItemViewHolder.java
@@ -38,7 +38,7 @@ class CategoryItemViewHolder extends CategoryPickerViewHolder {
 
     public interface CategoryItemViewHolderListener{
 
-        void selectItemCategory(int categoryId);
+        void selectItemCategory(long categoryId);
     }
 
 }
